rearrange uploading to S3
