+++ s3_client.py
@@ -1,0 +1,37 @@
+import io
+
+from minio import Minio
+import streamlit as st
+
+# Configuration de MinIO
+minio_client = Minio(
+    st.secrets['MINIO_INSTANCE_URL'],  # Remplacez par l'URL de votre instance MinIO
+    access_key= st.secrets['MINIO_ACCESS_KEY'],
+    secret_key=st.secrets["MINIO_SECRET_KEY"],
+    secure=False  # Choisissez True si vous utilisez HTTPS
+)
+
+# Nom du bucket sur MinIO
+bucket_name = st.secrets["MINIO_BUCKET_NAME"]
+
+def upload_to_minio(file):
+    # Lire le contenu du fichier téléchargé
+    file_content = file.read()
+
+    # Crée un flux en mémoire pour MinIO
+    file_stream = io.BytesIO(file_content)
+
+    # Nom du fichier (tu peux ajouter un timestamp pour éviter les collisions)
+    file_name = file.name
+
+    try:
+        # Upload sur MinIO
+        minio_client.put_object(
+            bucket_name,  # Le nom du bucket
+            file_name,    # Le nom du fichier sur MinIO
+            file_stream,  # Le flux du fichier
+            len(file_content)  # La taille du fichier
+        )
+        st.success(f"Fichier {file_name} téléchargé avec succès sur MinIO.")
+    except Exception as e:
+        st.error(f"Erreur lors de l'upload sur MinIO: {e}")
